Changes to migration and route
The get me route was changed to return a User if the request was done either by an admin or a normal user.
The migration for vehicules up, was changed to add a default picture without the uploads base dir.
The save path for vehicules picture patch route was change to be consistent with migration