Vehicule migration changed
The vehicule migration was change to spanish, to map better to the ui due to shared models in backend and frontend, also the backend feature of common was added.
All the vehicule related api endpoint were changed to use new model, also most of the api responses were changed to spanish.
A test migration was added to test some sqlx funtionality, this migration will be used for tests and will be removed before pushing to production.
The default picture for users was changed to a better one with black and white colors.